moved important buttons outside of tab
