+++ frontend/src/features/RS3/Player/Rs3PlayerInfo.tsx
@@ -10,9 +10,15 @@ import {
 	selectClanname,
 	selectQuestPoints,
 	selectIronStatus,
+	selectIsTracking,
 } from 'features/RS3/rs3Slice';
+import { selectAuthIsLoggedIn } from 'features/Common/commonSlice';
+
 import { Badge } from 'features/Common/Badge';
 import { Avatar } from 'features/RS3/Avatar';
+import TrackingButton from 'features/RS3/Player/Details/TrackingButton';
+import { FollowButton } from 'features/Common/FollowButton';
+
 import TotalXpIcon from 'assets/skillIcons/1_overall.png';
 import RunescoreIcon from 'assets/images/RuneScore.png';
 import QuestIcon from 'assets/images/questIcon.png';
@@ -26,47 +32,61 @@ export default function Rs3PlayerInfo() {
 	const badges = useAppSelector(selectBadges);
 	const questPoints = useAppSelector(selectQuestPoints);
 	const ironStatus = useAppSelector(selectIronStatus);
+	const isTracking = useAppSelector(selectIsTracking);
+	const isLoggedIn = useAppSelector(selectAuthIsLoggedIn);
+
+	const showFollowButton = isTracking && isLoggedIn && username;
 
 	return (
-		<div className="container--player-info p-d-flex p-ai-center p-flex-wrap">
-			<div className="container--avatar p-my-2">
-				<Avatar username={username} />
-			</div>
-			<div className="container--text-info">
-				<div className="p-d-flex p-ai-center">
-					<h1 className="p-mr-2 p-my-0">{username}</h1>
-					<IronIcon accountType={ironStatus} />
+		<div className="container--player-info p-d-flex p-ai-center p-flex-wrap p-jc-between">
+			<div className="p-d-flex p-ai-center p-flex-wrap">
+				<div className="container--avatar p-my-2">
+					<Avatar username={username} />
 				</div>
-				<p className="p-m-0">{clanname}</p>
+				<div className="container--text-info">
+					<div className="p-d-flex p-ai-center">
+						<h1 className="p-mr-2 p-my-0">{username}</h1>
+						<IronIcon accountType={ironStatus} />
+					</div>
+					<p className="p-m-0">{clanname}</p>
+				</div>
+				<div className="container--totalxp p-ml-5">
+					<img
+						className="icon--player-info p-mr-2"
+						src={TotalXpIcon}
+						alt="total xp"
+					/>
+					<span>{totalXp.toLocaleString()}</span>
+				</div>
+				<div className="container--runescore p-ml-5">
+					<img
+						className="icon--player-info p-mr-2"
+						src={RunescoreIcon}
+						alt="total xp"
+					/>
+					<span>{runescore.toLocaleString()}</span>
+				</div>
+				<div className="container--qp p-ml-5">
+					<img
+						className="icon--player-info p-mr-2"
+						src={QuestIcon}
+						alt="quest points"
+					/>
+					<span>{questPoints}</span>
+				</div>
+				<div className="container--badges p-ml-5">
+					{badges.map((b) => (
+						<Badge key={b} badge={b} />
+					))}
+				</div>
 			</div>
-			<div className="container--totalxp p-ml-5">
-				<img
-					className="icon--player-info p-mr-2"
-					src={TotalXpIcon}
-					alt="total xp"
-				/>
-				<span>{totalXp.toLocaleString()}</span>
-			</div>
-			<div className="container--runescore p-ml-5">
-				<img
-					className="icon--player-info p-mr-2"
-					src={RunescoreIcon}
-					alt="total xp"
-				/>
-				<span>{runescore.toLocaleString()}</span>
-			</div>
-			<div className="container--qp p-ml-5">
-				<img
-					className="icon--player-info p-mr-2"
-					src={QuestIcon}
-					alt="quest points"
-				/>
-				<span>{questPoints}</span>
-			</div>
-			<div className="container--badges p-ml-5">
-				{badges.map((b) => (
-					<Badge key={b} badge={b} />
-				))}
+			<div className="p-d-flex p-ai-center p-flex-wrap">
+				<TrackingButton />
+				{showFollowButton ? (
+					<FollowButton playerToFollow={username} />
+				) : (
+					<div />
+				)}
 			</div>
 		</div>
 	);
